chore: rename lib/common/common.h => lib/common/Common.h
