+++ src/lib/arch/Arch.h
@@ -24,7 +24,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 #if SYSAPI_WIN32
