+++ src/lib/arch/XArch.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 #include "common/stdexcept.h"
 #include <string>
