+++ src/lib/arch/multibyte.h
@@ -8,7 +8,7 @@
 #pragma once
 
 #include "arch/Arch.h"
-#include "common/common.h"
+#include "common/Common.h"
 
 #include <climits>
 #include <cstdlib>
