+++ src/lib/base/Event.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 #include "common/stdmap.h"
 
 class EventData
