+++ src/lib/base/Log.h
@@ -9,7 +9,7 @@
 
 #include "arch/Arch.h"
 #include "arch/IArchMultithread.h"
-#include "common/common.h"
+#include "common/Common.h"
 #include "common/stdlist.h"
 
 #include <stdarg.h>
