+++ src/lib/base/Stopwatch.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 //! A timer class
 /*!
