+++ src/lib/base/String.h
@@ -8,7 +8,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 #include <string>
 
 #include <stdarg.h>
