+++ src/lib/client/HelloBack.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 #include "deskflow/ProtocolTypes.h"
 #include "io/IStream.h"
