+++ src/lib/deskflow/Chunk.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 #include <base/EventTypes.h>
 
 class Chunk : public EventData
