+++ src/lib/deskflow/ClipboardChunk.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 #include "deskflow/Chunk.h"
 #include "deskflow/ClipboardTypes.h"
