+++ src/lib/deskflow/DaemonApp.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 #include <string>
