+++ src/lib/deskflow/KeyTypes.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 //! Key ID
 /*!
