+++ src/lib/mt/Lock.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 class Mutex;
 class CondVarBase;
