+++ src/lib/platform/OSXDragSimulator.h
@@ -6,7 +6,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 #import <CoreFoundation/CoreFoundation.h>
