+++ src/lib/platform/OSXScreenSaverUtil.h
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include "common/common.h"
+#include "common/Common.h"
 
 #if defined(__cplusplus)
 extern "C"
